Add parsing arguments from console
